requirements.txt
aiohttp==3.7.2
pyyaml==5.3.1
sqlalchemy==1.3.20
psycopg2==2.8.6
bcrypt==3.2.0
gino==1.0.1
pyjwt==1.7.1